refactor(backend): extract amp mode detection
